split up the input specs and sort the parser base class.
