+++ roll_witch/dice_bot/input/__init__.py
@@ -1,9 +1,13 @@
-from .regex import RegexInputParser, get_regex_parser
-from .token import TokenInputParser, get_token_parser
+from .parser.regex import RegexInputParser
+from .parser.token import TokenInputParser
+from .simple_regex import get_regex_parser
+from .token import get_token_parser
 
 __all__ = [
     "RegexInputParser",
     "TokenInputParser",
     "get_regex_parser",
-    "get_token_parser",
+    "get_token_parser"
 ]
+
+
