+++ roll_witch/dice_bot/input/spec/__init__.py
@@ -1,0 +1,6 @@
+from .base import InputPartSpec
+from .target import TargetSpec, TargetWithDiceSpec
+from .dice import DiceSpec, DiceWithModifierSpec
+from .modifier import ModifierSpec
+
+__all__ = ["InputPartSpec", "TargetSpec", "TargetWithDiceSpec", "DiceSpec", "DiceWithModifierSpec", "ModifierSpec"]
