+++ roll_witch/dice_bot/input/spec/base.py
@@ -1,0 +1,14 @@
+from typing import Pattern, Match
+
+from roll_witch.dice_bot.spec import RollSpec
+
+
+class InputPartSpec():
+    regex: Pattern
+    name: str
+
+    def matches_pattern(self, part_string) -> Match:
+        return self.regex.fullmatch(part_string)
+
+    def apply(self, target_match: Match) -> [RollSpec, None]:
+        return None
